Clean up 501(c)(3) status, hack night description (#110)
Copyedited the homepage. Moved the EIN down to the section on sponsors,
which is where people would more likely look for this information.
Updated the events section to remove references to Meetup and COVID-19,
since not all of our meetings are virtual anymore. Noted that some teams
like [OSM](https://github.com/codeforsanjose/OSM-SouthBay/) are also
meeting on a different schedule.