Migrate to Jekyll (#113)
Refactored the structure of every page into a shared layout with some
nested includes. Refactored basic information about the organization
into sitewide configuration variables.

This is the bare minimum needed to migrate to Jekyll, without taking
advantage of many of its features, for example, stable links,
collections (a more manageable alternative to the project information
loaded from a CSV at runtime), and reusable plugins. There’s also a lot
of outdated information, broken links, and typos that I didn’t touch. If
we decide to keep this site largely as it is, we can still optimize it
further in subsequent PRs.

Fixes #95 and fixes #111 and fixes #114.